nvapi,nvofapi: Slightly reorganize includes
Introduce `nvofapi_private.h`.
Move all external COM interface to new `interfaces` folder.
Use relative includes. This makes adding tests easier.
