+++ src/d3d12/nvofapi_d3d12_instance.h
@@ -23,7 +23,8 @@
 
 #pragma once
 
-#include "nvofapi/nvofapi_instance.h"
+#include "../nvofapi_private.h"
+#include "../nvofapi/nvofapi_instance.h"
 
 namespace dxvk {
     class NvOFInstanceD3D12 : public NvOFInstance {
