+++ src/nvofapi_d3d12.cpp
@@ -21,10 +21,9 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-#include "nvofapi/nvofapi_entrypoints.h"
 #include "nvofapi/nvofapi_image.h"
 #include "d3d12/nvofapi_d3d12_instance.h"
-
+#include "nvofapi_entrypoints.h"
 #include "util/util_env.h"
 #include "util/util_log.h"
 #include "util/util_string.h"
