+++ src/nvofapi_private.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "./nvapi_private.h"
+#include "../inc/nvofapi/nvOpticalFlowD3D12.h"
+#include "../inc/nvofapi/nvOpticalFlowVulkan.h"
