+++ src/resource_factory.cpp
@@ -1,5 +1,5 @@
 #include "resource_factory.h"
-#include "dxvk/dxvk_interfaces.h"
+#include "interfaces/dxvk_interfaces.h"
 #include "util/util_string.h"
 #include "util/util_log.h"
