+++ src/sysinfo/nvapi_adapter.cpp
@@ -1,5 +1,5 @@
 #include "nvapi_adapter.h"
-#include "../dxvk/dxvk_interfaces.h"
+#include "../interfaces/dxvk_interfaces.h"
 #include "../util/util_string.h"
 #include "../util/util_env.h"
 #include "../util/util_log.h"
